Final amorin

+++ appReservacion/src/main/java/com/bus/sistema/app_reservacion/ModSeguridad/Services/GenericService.java
@@ -10,11 +10,11 @@ public interface GenericService<T extends Object> {
 
     void delete(T entity);
 
-    void delete(Long id);
+    void delete(int id);
 
     void deleteInBatch(List<T> entities);
 
-    T find(Long id);
+    T find(int id);
 
     List<T> findAll();
 }
